simple_example/requirements.txt
Flask==1.1.2
opentelemetry-api==0.6b0
opentelemetry-ext-jaeger==0.6b0
opentelemetry-sdk==0.6b0
opentelemetry-ext-http-requests==0.6b0
opentelemetry-auto-instrumentation==0.6b0
opentelemetry-ext-otcollector==0.6b0
requests==2.23.0
